Update README.md
Displaying a clickable word LINK leading he user to the SOLIDITY website, to bring traffic. 
SoLiDiTy logo displayed as a user friendly link also leading individuals to the contribution section of the Solidity website. 
Size of the logo could be changed, this is an attribute I want to learn. 
I am assuming this is in c#, being this is a static. 
Can someone help me resize this LOGO without using HTML tags?